# perf: proto.Value remove interface{} overhead